Fix C4244 narrowing warnings via type promotion
Promote function signatures and local variables to wider types to eliminate
MSVC C4244 warnings without losing precision:

- dwgbuffer: numRemainingBytes() -> dint64, moveBitPos(dint64),
  crc8() start/end -> dint64, getUCSStr/get16bitStr size -> duint32
- dwgreader: maxPos -> duint64, startPos -> dint64,
  dwgSectionInfo compressed/encrypted -> duint64, pages map key -> duint64
- dwgutil: compressedSize/decompSize/compressedPos/decompPos -> duint64,
  compressedByte/decompByte index -> duint64, decompress18 locals -> duint64,
  decrypt18Hdr loop vars -> duint64, twoByteOffset litCount param -> duint64
- drw_entities/drw_objects: ms -> static_cast<duint32>(buf->size()),
  numRemainingBytes local -> dint64
- dxfwriter: writeInt64 byte extraction -> static_cast<char>
- libdwgr/dwgreader15/18/21: static_cast<size_t> at allocator boundaries

Use static_cast only where narrowing is intentional per DWG spec
(e.g., 32-bit XOR mask, allocator boundaries on 32-bit platforms).

+++ src/intern/dwgreader15.cpp
@@ -120,7 +120,7 @@ bool dwgReader15::readDwgHeader(DRW_Header& hdr){
         return false;
     if (!fileBuf->setPosition(si.address))
         return false;
-    std::vector<duint8> tmpByteStr(si.size);
+    std::vector<duint8> tmpByteStr(static_cast<size_t>(si.size));
     fileBuf->getBytes(tmpByteStr.data(), si.size);
     dwgBuffer buff(tmpByteStr.data(), si.size, &decoder);
     DRW_DBG("Header section sentinel= ");
